+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes qsf-integration project will be documented in this file.
 - implement sort rules
 - add minTokenWeight for Spellcheck
 - implement a logic to track the reason for a changed query
+- implement a include for the json builder
 
 ### Removed